Rework and simplify vendor

+++ build.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
 #
-# Android AOSP/MyAOSP/AOSPA/CM/SLIM/MAHDI build script
-# Version 2.0.12
+# Android AOSP/MyAOSP/AOSPA/CM/SLIM/MAHDI/EOS build script
+# Version 2.0.13
 #
 # https://source.android.com/source/initializing.html
 #
@@ -37,6 +37,8 @@ txtrst=$(tput sgr0)             # reset
 : ${CCACHE_NOSTATS:="false"}
 : ${CCACHE_DIR:="$(dirname $OUT)/ccache"}
 : ${THREADS:="$(cat /proc/cpuinfo | grep "^processor" | wc -l)"}
+#: ${STOCK_SYSTEM_IMAGE:="$(dirname $OUT)/../system.img"}
+: ${ODEX_MODE:="true"}
 
 # If there is more than one jdk installed, use latest 7.x
 if [ "`update-alternatives --list javac | wc -l`" -gt 1 ]; then
@@ -53,7 +55,6 @@ if [ "`update-alternatives --list javac | wc -l`" -gt 1 ]; then
 fi
 JVER=$(javac -version  2>&1 | head -n1 | cut -f2 -d' ')
 
-
 # Import command line parameters
 DEVICE="$1"
 EXTRAS="$2"
@@ -62,46 +63,55 @@ EXTRAS="$2"
 : ${DEVICE:="hammerhead"}
 
 # Get build version
-if [ -r ${DIR}/vendor/pa/vendor.mk ]; then
+if [ -r "${DIR}/vendor/pa/vendor.mk" ]; then
         VENDOR="aospa"
         VENDOR_LUNCH="pa_"
-        MAJOR=$(cat ${DIR}/vendor/pa/vendor.mk | grep 'ROM_VERSION_MAJOR := *' | sed  's/ROM_VERSION_MAJOR := //g')
-        MINOR=$(cat ${DIR}/vendor/pa/vendor.mk | grep 'ROM_VERSION_MINOR := *' | sed  's/ROM_VERSION_MINOR := //g')
-        MAINT=$(cat ${DIR}/vendor/pa/vendor.mk | grep 'ROM_VERSION_MAINTENANCE := *' | sed  's/ROM_VERSION_MAINTENANCE := //g')
-elif [ -r ${DIR}/vendor/slim/vendorsetup.sh ]; then
+        VENDOR_MK="${DIR}/vendor/pa/vendor.mk"
+elif [ -r "${DIR}/vendor/slim/config/common.mk" ]; then
         VENDOR="slim"
         VENDOR_LUNCH="slim_"
-        MAJOR=$(cat ${DIR}/vendor/slim/config/common.mk | grep 'PRODUCT_VERSION_MAJOR = *' | sed  's/PRODUCT_VERSION_MAJOR = //g')
-        MINOR=$(cat ${DIR}/vendor/slim/config/common.mk | grep 'PRODUCT_VERSION_MINOR = *' | sed  's/PRODUCT_VERSION_MINOR = //g')
-        MAINT=$(cat ${DIR}/vendor/slim/config/common.mk | grep 'PRODUCT_VERSION_MAINTENANCE = *' | sed  's/PRODUCT_VERSION_MAINTENANCE = //g')
-elif [ -r ${DIR}/vendor/cm/config/common.mk ]; then
+        VENDOR_MK="${DIR}/vendor/slim/config/common.mk"
+elif [ -r "${DIR}/vendor/cm/config/common.mk" ]; then
         VENDOR="cm"
         VENDOR_LUNCH=""
-        MAJOR=$(cat ${DIR}/vendor/cm/config/common.mk | grep 'PRODUCT_VERSION_MAJOR = *' | sed  's/PRODUCT_VERSION_MAJOR = //g')
-        MINOR=$(cat ${DIR}/vendor/cm/config/common.mk | grep 'PRODUCT_VERSION_MINOR = *' | sed  's/PRODUCT_VERSION_MINOR = //g')
-        MAINT=$(cat ${DIR}/vendor/cm/config/common.mk | grep 'PRODUCT_VERSION_MAINTENANCE = *' | sed  's/PRODUCT_VERSION_MAINTENANCE = //g')
-elif [ -r ${DIR}/vendor/mahdi/config/common.mk ]; then
+        VENDOR_MK="${DIR}/vendor/cm/config/common.mk"
+        VENDOR_PB_XML_URL="https://gist.github.com/eugenesan/857b32fb134cbe26fe99/raw/proprietary_vendor_lge.xml"
+        VENDOR_PB_XML=".repo/local_manifests"
+elif [ -r "${DIR}/vendor/mahdi/config/common_versions.mk" ]; then
         VENDOR="mahdi"
         VENDOR_LUNCH=""
-        MAJOR=$(cat ${DIR}/vendor/mahdi/config/common_versions.mk | grep 'PRODUCT_VERSION_MAJOR = *' | sed  's/PRODUCT_VERSION_MAJOR = //g')
-        MINOR=$(cat ${DIR}/vendor/mahdi/config/common_versions.mk | grep 'PRODUCT_VERSION_MINOR = *' | sed  's/PRODUCT_VERSION_MINOR = //g')
-        MAINT=$(cat ${DIR}/vendor/mahdi/config/common_versions.mk | grep 'PRODUCT_VERSION_MAINTENANCE = *' | sed  's/PRODUCT_VERSION_MAINTENANCE = //g')
-elif [ -r ${DIR}/vendor/myaosp/configs/common.mk ]; then
+        VENDOR_MK="${DIR}/vendor/mahdi/config/common_versions.mk"
+elif [ -r "${DIR}/vendor/eos/config/common_versions.mk" ]; then
+        VENDOR="eos"
+        VENDOR_LUNCH=""
+        VENDOR_MK="${DIR}/vendor/eos/config/common_versions.mk"
+elif [ -r "${DIR}/vendor/myaosp/product/common.mk" ]; then
         VENDOR="myaosp"
-        VENDOR_LUNCH="myaosp_"
-        MAJOR=$(cat ${DIR}/build/core/version_defaults.mk | grep 'PLATFORM_VERSION := *' | awk '{print $3}' | cut -f2 -d= | cut -f1 -d.)
-        MINOR=$(cat ${DIR}/build/core/version_defaults.mk | grep 'PLATFORM_VERSION := *' | awk '{print $3}' | cut -f2 -d= | cut -f2 -d.)
-        MAINT=$(cat ${DIR}/build/core/version_defaults.mk | grep 'PLATFORM_VERSION := *' | awk '{print $3}' | cut -f2 -d= | cut -f3 -d.)
-elif [ -r ${DIR}/build/core/version_defaults.mk ]; then
+        VENDOR_LUNCH=""
+        VENDOR_MK="${DIR}/build/core/version_defaults.mk"
+elif [ -r "${DIR}/build/core/version_defaults.mk" ]; then
         VENDOR="aosp"
         VENDOR_LUNCH="full_"
-        MAJOR=$(cat ${DIR}/build/core/version_defaults.mk | grep 'PLATFORM_VERSION := *' | awk '{print $3}' | cut -f2 -d= | cut -f1 -d.)
-        MINOR=$(cat ${DIR}/build/core/version_defaults.mk | grep 'PLATFORM_VERSION := *' | awk '{print $3}' | cut -f2 -d= | cut -f2 -d.)
-        MAINT=$(cat ${DIR}/build/core/version_defaults.mk | grep 'PLATFORM_VERSION := *' | awk '{print $3}' | cut -f2 -d= | cut -f3 -d.)
+        VENDOR_MK="${DIR}/build/core/version_defaults.mk"
 else
         echo -e "${redbld}Invalid android tree, exiting...${txtrst}"
         exit 1
 fi
+
+if [ "${VENDOR}" == "aosp" ] || [ "${VENDOR}" == "myaosp" ]; then
+        MAJOR=$(cat ${VENDOR_MK} | grep 'PLATFORM_VERSION := *' | awk '{print $3}' | cut -f2 -d= | cut -f1 -d.)
+        MINOR=$(cat ${VENDOR_MK} | grep 'PLATFORM_VERSION := *' | awk '{print $3}' | cut -f2 -d= | cut -f2 -d.)
+        MAINT=$(cat ${VENDOR_MK} | grep 'PLATFORM_VERSION := *' | awk '{print $3}' | cut -f2 -d= | cut -f3 -d.)
+elif [ "${VENDOR}" == "aospa" ]; then
+        MAJOR=$(cat ${VENDOR_MK} | grep 'ROM_VERSION_MAJOR = *' | grep -v "^#" | sed  's/ROM_VERSION_MAJOR = //g')
+        MINOR=$(cat ${VENDOR_MK} | grep 'ROM_VERSION_MINOR = *' | grep -v "^#" | sed  's/ROM_VERSION_MINOR = //g')
+        MAINT=$(cat ${VENDOR_MK} | grep 'ROM_VERSION_MAINTENANCE = *' | grep -v "^#" | sed  's/ROM_VERSION_MAINTENANCE = //g')
+else
+        MAJOR=$(cat ${VENDOR_MK} | grep 'PRODUCT_VERSION_MAJOR = *' | grep -v "^#" | sed  's/PRODUCT_VERSION_MAJOR = //g')
+        MINOR=$(cat ${VENDOR_MK} | grep 'PRODUCT_VERSION_MINOR = *' | grep -v "^#" | sed  's/PRODUCT_VERSION_MINOR = //g')
+        MAINT=$(cat ${VENDOR_MK} | grep 'PRODUCT_VERSION_MAINTENANCE = *' | grep -v "^#" | sed  's/PRODUCT_VERSION_MAINTENANCE = //g')
+fi
+
 VERSION=$VENDOR-$MAJOR.$MINOR$([ -n "$MAINT" ] && echo .)$MAINT
 
 # If there is no extra parameter, reduce parameters index by 1
@@ -122,10 +132,12 @@ if [ "${USE_CCACHE}" == "true" ]; then
         CCACHE="$(which ccache)"
         if [ -n "${CCACHE}" ]; then
                 export CCACHE_COMPRESS="1"
+                export USE_SYSTEM_CCACHE="1"
                 echo -e "${bldblu}Using system ccache with compression enabled [${CCACHE}]${txtrst}"
         elif [ -r "${DIR}/prebuilts/misc/linux-x86/ccache/ccache" ]; then
                 CCACHE="${DIR}/prebuilts/misc/linux-x86/ccache/ccache"
                 echo -e "${bldblu}Using prebuilt ccache [${CCACHE}]${txtrst}"
+                unset USE_SYSTEM_CCACHE
         else
                 echo -e "${bldblu}No suitable ccache found, disabling ccache usage${txtrst}"
                 unset USE_CCACHE
@@ -183,6 +195,17 @@ case "$EXTRAS" in
                 THREADS=$threads
         ;;
         clean|cclean)
+		# Un-Mount stock image
+		if [ $(mount | grep "${DIR}/out/system.ext4" | grep -c "${DIR}/out/stock/system") == 1 ]; then
+			echo "${bldblu}Unmounting [${DIR}/out/stock/system]${txtrst}"
+			sudo umount "${DIR}/out/stock/system"
+		fi
+
+		if [ $(mount | grep "${DIR}/out/system.ext4" | grep -c "${DIR}/out/stock/system") != 0 ]; then
+			echo "${bldred}Unmounting failed [${DIR}/out/stock/system]${txtrst}"
+			exit 1
+		fi
+
                 echo -e "${bldblu}Cleaning intermediates and output files${txtrst}"
                 export CLEAN_BUILD="true"
                 [ -d "${DIR}/out" ] && rm -Rf ${DIR}/out/*
@@ -230,19 +253,79 @@ else
         echo -e "${bldcya}No blobs/prebuilts script in this tree${txtrst}"
 fi
 
-# Export global variables for buildsystem
+# Prepare stock image
+if [ -r "${STOCK_SYSTEM_IMAGE}" ]; then
+	echo -en "${bldblu}Analyzing stock image [${STOCK_SYSTEM_IMAGE}]:${txtrst}"
+	SB=`od -A n -h -j 0 -N 6 ${STOCK_SYSTEM_IMAGE} | sed 's/ //g'`
+	echo "${SB}"
+
+	if [ ! -r "${DIR}/out/system.ext4" ]; then
+		mkdir -p "${DIR}/out"
+
+		if [ "${SB}" == "000000000000" ]; then
+			echo -e "${bldblu}Copying image as already in ext4 format${txtrst}"
+			cp "${STOCK_SYSTEM_IMAGE}" "${DIR}/out/system.ext4"
+		else
+			if [ ! -x "${DIR}/out/simg2img" ]; then
+				echo -e "${bldblu}Preparing to compile simg2img from source${txtrst}"
+				[ -n `which gcc` ] || NEED_PACKAGES="${NEED_PACKAGES} build-essential"
+				[ -n `which git` ] || NEED_PACKAGES="${NEED_PACKAGES} git"
+				[ -r /usr/include/zlib.h ] || NEED_PACKAGES="${NEED_PACKAGES} zlib1g-dev"
+				if [ -n "${NEED_PACKAGES}" ]; then
+					echo -e "${bldcya}You need install additional packages:\n$ sudo aptitude install ${NEED_PACKAGES}${txtrst}"
+					exit 1
+				fi
+
+				echo -e "${bldblu}Compiling simg2img tool${txtrst}"
+				cd ${DIR}/system/core/libsparse
+				gcc -o ${DIR}/out/simg2img simg2img.c sparse*.c backed_block.c output_file.c -lz -Iinclude
+				cd - > /dev/null
+			fi
+
+			echo -e "${bldblu}Converting [${STOCK_SYSTEM_IMAGE}] to ext4 format${txtrst}"
+			${DIR}/out/simg2img "${STOCK_SYSTEM_IMAGE}" "${DIR}/out/system.ext4" > /dev/null
+		fi
+	else
+                echo -e "${bldblu}[${DIR}/out/system.ext4] already exists${txtrst}"
+	fi
+
+	# Mount stock image
+	if [ $(mount | grep "${DIR}/out/system.ext4" | grep -c "${DIR}/out/stock/system") == 1 ]; then
+                echo -e "${bldblu}[${DIR}/out/stock/system] is already mounted${txtrst}"
+	else
+                echo -e "${bldblu}Mounting stock image${txtrst}"
+		mkdir -p "${DIR}/out/stock/system"
+
+		# For ext4 (To prevent hanging when unmounting, use: ro,noexec,noload)
+		sudo mount -o loop,ro,noexec,noload "${DIR}/out/system.ext4" "${DIR}/out/stock/system"
+	fi
+
+	# Inject GAPPS (TODO: Validate config location for all ROMs)
+#	if [ -r "${DIR}/out/stock/system/build.prop" ] && ! grep -q "common_gapps" "vendor/${VENDOR}/config/common.mk"; then
+#		echo -e "${bldblu}Injecting google APPS${txtrst}"
+#		echo -e "# Include common stock google apps\n-include vendor/${VENDOR}/config/common_gapps.mk" >> "vendor/${VENDOR}/config/common.mk"
+#	else
+#		echo -e "${bldcya}Google APPS already injected or something wrong with stock image [${STOCK_SYSTEM_IMAGE}]${txtrst}"
+#	fi
+else
+	echo -e "${bldcya}No stock image provided at [${STOCK_SYSTEM_IMAGE}]${txtrst}"
+fi
+
+# Export global variables for build system
 export PREFS_FROM_SOURCE
 export WITHOUT_PROP_APPS
 export WITHOUT_MC
+#export DISABLE_DEXPREOPT=false
+#export WITH_DEXPREOPT=true
 
 # Decide if we enter interactive mode or default build mode
 if [ -n "${INTERACTIVE}" ]; then
         echo -e "\n${bldblu}Enabling interactive mode. Possible commands are:${txtrst}"
 
-        if [ "${VENDOR}" == "cm" ] || [ "${VENDOR}" == "mahdi" ]; then
+        if [ "${VENDOR}" == "cm" ] || [ "${VENDOR}" == "mahdi" ] || [ "${VENDOR}" == "eos" ]; then
                 echo -e "To prepare device environment:[${bldgrn}breakfast ${VENDOR_LUNCH}${DEVICE}${txtrst}]"
         else
-                echo -e "To prepare device environment:[${bldgrn}lunch ${VENDOR_LUNCH}${DEVICE}-userdebug${txtrst}]"
+                echo -e "To prepare device environment:[${bldgrn}lunch ${VENDOR_LUNCH}${DEVICE}-user${txtrst}]"
         fi
 
         if [ "${VENDOR}" == "aosp" ]; then
@@ -260,22 +343,27 @@ if [ -n "${INTERACTIVE}" ]; then
         echo -e "${bldblu}Dropping to interactive shell...${txtrst}"
         bash --init-file ${DIR}/build/envsetup.sh -i
 else
+	# Optionally switch to ODEX modes
+	#if [ "${ODEX_MODE}" == "true" ]; then
+	#	sed -i 's/combo-userdebug/combo-user/' "vendor/${VENDOR}/vendorsetup.sh"
+	#fi
+
         # Setup environment
         echo -e "\n${bldblu}Setting up environment${txtrst}"
         . build/envsetup.sh
 
         # Preparing
         echo -e "\n${bldblu}Preparing device [${DEVICE}]${txtrst}"
-        if [ "${VENDOR}" == "cm" ] || [ "${VENDOR}" == "mahdi" ]; then
+        if [ "${VENDOR}" == "cm" ] || [ "${VENDOR}" == "mahdi" ] || [ "${VENDOR}" == "eos" ]; then
                 breakfast "${VENDOR_LUNCH}${DEVICE}"
         else
-                lunch "${VENDOR_LUNCH}${DEVICE}-userdebug"
+                lunch "${VENDOR_LUNCH}${DEVICE}-user"
         fi
 
         echo -e "${bldblu}Starting compilation${txtrst}"
         if [ "${VENDOR}" == "aosp" ] || [ "${VENDOR}" == "myaosp" ]; then
                 schedtool -B -n 1 -e ionice -n 1 make -j${THREADS} ${CCACHE_OPT} ${JAVA_VERSION} otapackage
-        elif [ "${VENDOR}" == "cm" ] || [ "${VENDOR}" == "mahdi" ]; then
+        elif [ "${VENDOR}" == "cm" ] || [ "${VENDOR}" == "mahdi" ] || [ "${VENDOR}" == "eos" ]; then
                 brunch "${VENDOR_LUNCH}${DEVICE}"
         else
                 mka bacon
